fakevtctlclient: Factor out the core logic such that it can be reused for a fake of vtworkerclient.
