+++ go/vt/vtctl/fakevtctlclient/fake_loggerevent_streamingclient_test.go
@@ -2,16 +2,12 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// Package fakevtctlclient implements the vtctlclient interface and is used as mock in unit tests.
 package fakevtctlclient
 
 import (
 	"errors"
 	"strings"
 	"testing"
-	"time"
-
-	"golang.org/x/net/context"
 )
 
 func TestStreamOutputAndError(t *testing.T) {
@@ -23,7 +19,7 @@ func TestStreamOutput(t *testing.T) {
 }
 
 func verifyStreamOutputAndError(t *testing.T, wantErr error) {
-	fake := NewFakeVtctlClient()
+	fake := NewFakeLoggerEventStreamingClient()
 	args := []string{"CopySchemaShard", "test_keyspace/0", "test_keyspace/2"}
 	output := []string{"event1", "event2"}
 	err := fake.RegisterResult(args,
@@ -33,7 +29,7 @@ func verifyStreamOutputAndError(t *testing.T, wantErr error) {
 		t.Fatal(err)
 	}
 
-	stream, errFunc := fake.ExecuteVtctlCommand(context.Background(), args, time.Hour, 10*time.Second)
+	stream, errFunc := fake.StreamResult(args)
 
 	// Verify output and error.
 	i := 0
@@ -56,8 +52,8 @@ func verifyStreamOutputAndError(t *testing.T, wantErr error) {
 }
 
 func TestNoResultRegistered(t *testing.T) {
-	fake := NewFakeVtctlClient()
-	stream, errFunc := fake.ExecuteVtctlCommand(context.Background(), []string{"ListShardTablets", "test_keyspace/0"}, time.Hour, 10*time.Second)
+	fake := NewFakeLoggerEventStreamingClient()
+	stream, errFunc := fake.StreamResult([]string{"ListShardTablets", "test_keyspace/0"})
 	if stream != nil {
 		t.Fatalf("No stream should have been returned because no matching result is registered.")
 	}
@@ -71,7 +67,7 @@ func TestNoResultRegistered(t *testing.T) {
 }
 
 func TestResultAlreadyRegistered(t *testing.T) {
-	fake := NewFakeVtctlClient()
+	fake := NewFakeLoggerEventStreamingClient()
 	errFirst := fake.RegisterResult([]string{"ListShardTablets", "test_keyspace/0"}, "output1", nil)
 	if errFirst != nil {
 		t.Fatalf("Registering the result should have been successful. Error: %v", errFirst)
